refactor(DriveFileManager): Moved offline availability code to the proper extension
